// src/graphqlschema.d.ts
export type ID = string;
// ...
export interface IRepository {
  __typename: "Repository";
  id: ID;
  name: string;

  /**
   * @deprecated "use name instead"
   */
  uri: string;
  description: string;
  language: string;
  enabled: boolean;
  createdAt: string;
  updatedAt: string | null;
  commit: IGitCommit | null;
  mirrorInfo: IMirrorRepositoryInfo;
  externalRepository: IExternalRepository | null;

  /**
   * @deprecated "use Repository.mirrorInfo.cloneInProgress instead"
   */
  cloneInProgress: boolean;
  lastIndexedRevOrLatest: IGitCommit | null;
  textSearchIndex: IRepositoryTextSearchIndex | null;
  url: string;
  externalURLs: IExternalLink[];
  defaultBranch: IGitRef | null;
  gitRefs: IGitRefConnection;
  branches: IGitRefConnection;
  tags: IGitRefConnection;
  comparison: IRepositoryComparison;
  contributors: IRepositoryContributorConnection;
  symbols: ISymbolConnection;
  packages: IPackageConnection;
  dependencies: IDependencyConnection;
  listTotalRefs: ITotalRefList;
  redirectURL: string | null;
  viewerCanAdminister: boolean;
}
// ...
export interface IGitCommit {
  __typename: "GitCommit";
  id: ID;
  repository: IRepository;
  oid: any;
  abbreviatedOID: string;
  author: ISignature;
  committer: ISignature | null;
  message: string;
  subject: string;
  body: string | null;
  parents: IGitCommit[];
  url: string;
  canonicalURL: string;
  externalURLs: IExternalLink[];
  tree: IGitTree | null;
  blob: IGitBlob | null;
  file: File2 | null;
  languages: string[];
  ancestors: IGitCommitConnection;
  behindAhead: IBehindAheadCounts;
  symbols: ISymbolConnection;
  packages: IPackageConnection;
  dependencies: IDependencyConnection;
}
// ...
export interface IExternalLink {
  __typename: "ExternalLink";
  url: string;
  serviceType: string | null;
}
// ...
export interface ISymbolConnection {
  __typename: "SymbolConnection";
  nodes: ISymbol[];
  pageInfo: IPageInfo;
}
// ...
export interface IPackageConnection {
  __typename: "PackageConnection";
  nodes: IPackage[];
  totalCount: number;
  pageInfo: IPageInfo;
}
// ...
export interface IDependencyConnection {
  __typename: "DependencyConnection";
  nodes: IDependency[];
  totalCount: number;
  pageInfo: IPageInfo;
}

export interface IMirrorRepositoryInfo {
  __typename: "MirrorRepositoryInfo";
  remoteURL: string;
  cloneInProgress: boolean;
  cloneProgress: string | null;
  cloned: boolean;
  updatedAt: string | null;
}

export interface IExternalRepository {
  __typename: "ExternalRepository";
  id: string;
  serviceType: string;
  serviceID: string;
}

export interface IRepositoryTextSearchIndex {
  __typename: "RepositoryTextSearchIndex";
  repository: IRepository;
  status: IRepositoryTextSearchIndexStatus | null;
  refs: IRepositoryTextSearchIndexedRef[];
}
// ...
export interface IGitRef {
  __typename: "GitRef";
  id: ID;
  name: string;
  abbrevName: string;
  displayName: string;
  prefix: string;
  type: GitRefType;
  target: IGitObject;
  repository: IRepository;
  url: string;
}

export const enum GitRefType {
  GIT_BRANCH = "GIT_BRANCH",
  GIT_TAG = "GIT_TAG",
  GIT_REF_OTHER = "GIT_REF_OTHER"
}

export interface IGitObject {
  __typename: "GitObject";
  oid: any;
  abbreviatedOID: string;
  commit: IGitCommit | null;
  type: GitObjectType;
}

export const enum GitObjectType {
  GIT_COMMIT = "GIT_COMMIT",
  GIT_TAG = "GIT_TAG",
  GIT_TREE = "GIT_TREE",
  GIT_BLOB = "GIT_BLOB",
  GIT_UNKNOWN = "GIT_UNKNOWN"
}

export const enum GitRefOrder {
  AUTHORED_OR_COMMITTED_AT = "AUTHORED_OR_COMMITTED_AT"
}

export interface IGitRefConnection {
  __typename: "GitRefConnection";
  nodes: IGitRef[];
  totalCount: number;
  pageInfo: IPageInfo;
}

export interface IRepositoryComparison {
  __typename: "RepositoryComparison";
  range: IGitRevisionRange;
  commits: IGitCommitConnection;
  fileDiffs: IFileDiffConnection;
}
// ...
export interface IRepositoryContributorConnection {
  __typename: "RepositoryContributorConnection";
  nodes: IRepositoryContributor[];
  totalCount: number;
  pageInfo: IPageInfo;
}

export interface IRepositoryContributor {
  __typename: "RepositoryContributor";
  person: IPerson;
  count: number;
  repository: IRepository;
  commits: IGitCommitConnection;
}

export interface ICommitsOnRepositoryContributorArguments {
  first?: number | null;
}

export interface ITotalRefList {
  __typename: "TotalRefList";
  repositories: IRepository[];
  total: number;
}
// ...
export interface IDiscussionThreadTargetRepo {
  __typename: "DiscussionThreadTargetRepo";
  repository: IRepository;
  path: string | null;
  branch: IGitRef | null;
  revision: IGitRef | null;
  selection: IDiscussionThreadTargetRepoSelection | null;
  relativePath: string | null;
  relativeSelection: IDiscussionSelectionRange | null;
}

export interface IRelativePathOnDiscussionThreadTargetRepoArguments {
  rev: string;
}

export interface IRelativeSelectionOnDiscussionThreadTargetRepoArguments {
  rev: string;
}

export interface IDiscussionThreadTargetRepoSelection {
  __typename: "DiscussionThreadTargetRepoSelection";
  startLine: number;
  startCharacter: number;
  endLine: number;
  endCharacter: number;
  linesBefore: string[];
  lines: string[];
  linesAfter: string[];
}

export interface IDiscussionSelectionRange {
  __typename: "DiscussionSelectionRange";
  startLine: number;
  startCharacter: number;
  endLine: number;
  endCharacter: number;
}